% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bycatch_mort}
\alias{bycatch_mort}
\title{Species bycatch and mortalities from sets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 157 rows and 5 columns.}
\usage{
bycatch_mort
}
\description{
A dataset species bycatch and accidental fish mortality attributed to seines conducted by the Juvenile Salmon Program
}
\details{
\describe{
 \item{seine_id}{Unique ID assigned to each seine conducted}
 \item{bm_species}{First two letters of the first two words of the common name of the fish. Example "TISC" = Tide-pool Sculpin. Was either a part of seine bycatch, or died during holding in net}
 \item{bm_status}{Whether the species was a bycatch or a mortality}
 \item{bm_ageclass}{The approximate age of the fish. Y = young-of-year, J = juvenile, A = adult}
 \item{bm_count}{The number of individuals of the species that was either bycatch or mortality}

}
}
\keyword{datasets}
