% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_dynamics_funcs.R
\name{cat_function}
\alias{cat_function}
\title{CPUE function}
\usage{
cat_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to calculate CPUE
}
\examples{
cats()
}
\keyword{cats}
