% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{display_hakai_colours}
\alias{display_hakai_colours}
\title{Display hcolors}
\usage{
display_hakai_colours(palette = c("hakai"))
}
\arguments{
\item{palette}{Character string indicating a palette of colors.}
}
\description{
displays the discrete hcolors palettes
}
\examples{
display_hakai_colours()
}
