% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration_dynamics_funcs.R
\name{.onAttach}
\alias{.onAttach}
\title{Cat function}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to detrmine if you are a cool person
}
\examples{
cat_function()
}
\keyword{cats}
