% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finclip_metadata}
\alias{finclip_metadata}
\title{Storage information for fin clip samples}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2264 rows and 12 columns.}
\usage{
finclip_metadata
}
\description{
Metadata associated with fin clip samples obtained from fish caught by the Juvenile Salmon Program, for use in genetic stock ID analysis
}
\details{
\describe{
 \item{ufn}{The alphanumeric string that uniquely identifies a fish. Its format is "U" followed by a sequence of numbers.}
 \item{sample_type}{The type of sample being collected.}
 \item{sample_id}{The ID of the cell in which the fin clip is placed. This enables results to be linked back to the sample and fish from which it was collected.}
 \item{container_id}{The string that uniquely identifies the Whatman sheet in which fin clips have been placed.}
 \item{container_cell}{Not applicable}
 \item{sample_location}{A concatenation of container_id and container_cell. Functionally, it is the same as container_id}
 \item{comments_sample}{Comments regarding notable characteristics/properties of the particular sample}
 \item{sample_status}{Results status of the sample}
 \item{sample_qc_flag}{Indicates whether there is a qc error for the sample's metadata (not quality of the actual sample)}
 \item{sample_flag_reason}{Describes why the sample has been flagged.}
 \item{sample_qc_log}{Log of changes to any sample metadata during qa/qc}

}
}
\keyword{datasets}
