% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_field_data}
\alias{fish_field_data}
\title{Biological measurements and metadata associated with fish caught in the field}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12884 rows and 12 columns.}
\usage{
fish_field_data
}
\description{
A dataset containing variables collected from field measurements of fish collected by the Juvenile Salmon Program
}
\details{
\describe{
 \item{ufn}{The alphanumeric string that uniquely identifies a fish. Joining to fish_lab data by UFN will filter out fish that haven't been dissected, because un-dissected fish do no have a UFN. To retain un-dissected fish when joining to fish_lab_data use SEMSP ID.}
 \item{semsp_id}{A concatenation of date, site, species, and order the fish was pulled out of the net was used to uniquely identify the fish. In 2017 this was method was abandonded, and instead we assigned seine IDs and UFNs to uniquely identify seines and fish}
 \item{seine_id}{Unique ID assigned to each seine conducted}
 \item{species}{The species of fish collected. SO = Sockeye, PI = Pink, CU = Chum, CO = Coho, HE = Herring, CK = Chinook}
 \item{package_id}{Unique identifier to identify the package (aka bundle) that a fish was grouped with in the field, and then subsequently stored with in the freezer until it was dissected}
 \item{fish_time_out}{Time that the fish was euthanized after being taken out of the seine. In conjunction with fis_time_dewar you an calculate the handling time of the fish before being flash frozen in liquid nitrogenh}
 \item{fish_dewar}{Time that the fish was put into the liquid nitrogen dry shipper. In conjunction with fish_time_dewar you an calculate the handling time of the fish before being flash frozen in liquid nitrogen}
 \item{fork_length_field}{The fork length of the fish measured while in the field using a 15 cm flexible ruler}
 \item{height_field}{The height of the fish measured perpendicular to the lateral line from the anterior insertion of the dorsal fin to the ventral side of the fish}
 \item{weight_field}{The weight of the fish measured in the field using a spring scale. Precision is +/- 1 g}
 \item{analysis_planned}{The type of analysis the fish is to undergo (or has underwent). "Field loused only" fish are not retained, and therefore have no SEMSP ID or package ID for inventory.}
 \item{dissection_status}{Whether or not the fish had been dissected at the time the data package was released}

}
}
\keyword{datasets}
