% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{hakai_palette}
\alias{hakai_palette}
\title{A Hakai Institute palette generator}
\usage{
hakai_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{hakai_web}, \code{Zissou1},  \code{Darjeeling1}}

\item{n}{Number of colors desired. Most palettes now only
have 4 or 5 colors.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
A Hakai Institute palette generator
}
\examples{
hakai_palette("hakai_web")


# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- hakai_palette(21, name = "Zissou1", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
