% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{hakai_palettes}
\alias{hakai_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
hakai_palettes
}
\description{
Use \code{\link{hakai_palette}} to construct palettes of desired length.
}
\keyword{datasets}
