% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{hcolours}
\alias{hcolours}
\title{Vectors of colors for figures}
\usage{
hcolours(palette = c("hakai", "pal2"))
}
\arguments{
\item{palette}{Character string indicating a palette of colors.}
}
\value{
Vector of character strings representing the chosen palette of colors.
}
\description{
Creates different vectors of related colors that may be useful for figures.
}
