% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_templates.R
\name{html_report_template}
\alias{html_report_template}
\title{An R Markdown .html report template}
\format{
R Markdown file with associated preset yaml front matter that defines a nice standard aesthetic for .html reports. This includes floating TOC, cosmo theme, code folding, and numbered sections.
}
\description{
This report template can be accessed by loading the hakaiR
package in R: `library(hakaiR)` and then click `New File > R Markdown` and
in the left side bar navigate down to `Templates` and select hakaiR.html report
}
