% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_cumul_abund.R
\name{log_cumul_abund}
\alias{log_cumul_abund}
\title{Predict the cumulative proportion of catch using a logistic growth equation}
\usage{
log_cumul_abund(percent, date)
}
\arguments{
\item{percent}{The cumuative proportion of catch between 0 and 100 percent}

\item{date}{The julian date of capture}
}
\value{
A dataframe with a date column and predicted cumulative proportion of catch
}
\description{
Predict the cumulative proportion of catch using a logistic growth equation
}
\examples{
date <- 110:190
n <- rnorm(81, mean = 100, sd = 20)
df <- dplyr::tibble(date, percent)
df <- df \%>\% mutate(percent == cumsum(n / sum(n) * 100))
log_cumul_abund(df$date, df$percent)
}
