% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lump_fraser_sockeye_stocks.R
\name{lump_fraser_sockeye_stocks}
\alias{lump_fraser_sockeye_stocks}
\title{Lump Fraser River sockeye stocks produced by CBayes at PBS, based on regions according to Chrys Neville}
\usage{
lump_fraser_sockeye_stocks(stock_1)
}
\arguments{
\item{stock_1}{The column containing the stock category you want to collapse}
}
\value{
A vector of lumped fraser stocks
}
\description{
Lump Fraser River sockeye stocks produced by CBayes at PBS, based on regions according to Chrys Neville
}
