% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_cond.R
\name{mutate_cond}
\alias{mutate_cond}
\title{A conditional mutate function that allows you to mutate a variable based on another}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data, }{the data frame or tibble you want to mutate a column in}

\item{x}{The column you want to test for a logical condition}

\item{y}{The coloumn you want to mutate based on logic from column x}
}
\description{
A conditional mutate function that allows you to mutate a variable based on another
}
\examples{
mutate_cond(df, x == 1, y == "c" )
}
\keyword{conditional}
\keyword{mutate}
