% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{package_data}
\alias{package_data}
\title{Inventory and other metadata associated with packages of fish caught from seines}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 552 rows and 13 columns.}
\usage{
package_data
}
\description{
A dataset containing variables describing the contents and location of packages of fish caught in Juvenile Salmon Program surveys
}
\details{
\describe{
\item{package_id}{The alphanumeric string that uniquely identifies a package of fish}
\item{seine_id}{Unique ID assigned to each seine conducted}
\item{pkg_species}{The species of fish within a package.}
\item{number_of_fish}{Number of fish within a package}
\item{field_liced_status}{Whether fish received lice enumeration in the field prior to packaging, or no lice enumeration has been performed}
\item{pkg_time_out}{Time the first fish of the package was taken out of the net and euthanized}
\item{pkg_time_dewar}{Time the full package was placed in the dewar for preservation}
\item{pkg_comments}{Notes regarding the contents or nature of the package}

}
}
\keyword{datasets}
