% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_templates.R
\name{pdf_report_template}
\alias{pdf_report_template}
\title{An R Markdown .pdf report template}
\format{
R Markdown file with associated .tex and .png files. Includes in the header: Hakai Instute logo, second level heading, default Times New Roman 12 pt font and some other LaTex code to make things look nice.
}
\description{
This report template can be accessed by loading the hakaiR
package in R: `library(hakaiR)` and then click `New File > R Markdown` and
in the left side bar navigate down to `Templates` and select hakaiR.pdf report
}
