% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_metadata}
\alias{rna_metadata}
\title{Storage information for RNA samples}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4173 rows and 12 columns.}
\usage{
rna_metadata
}
\description{
Metadata associated with tissue samples obtained from fish caught by the
Juvenile Salmon Program, for use in RNA:DNA ratio analysis and viral disease
states. The following tissue types are collected: muscle, gill, brain, spleen,
liver, heart, and kidney
}
\details{
\describe{
 \item{ufn}{The alphanumeric string that uniquely identifies a fish. Its format is "U" followed by a sequence of numbers.}
 \item{sample_type}{The type of sample being collected.}
 \item{sample_id}{The sample_location value (detailed below). This enables results to be linked back to the sample and fish from which it was collected.}
 \item{container_id}{The string that uniquely identifies the tray in which sets of sample vials are stored.}
 \item{container_cell}{The row (A-H) and column (1-12) in which a sample is stored in its container}
 \item{sample_location}{A concatenation of container_id and container_cell, which forms a unique location code for the purposes of identifying potential duplicate entries during qa/qc.}
 \item{comments_sample}{Comments regarding notable characteristics/properties of the particular sample}
 \item{sample_status}{Results status of the sample}
 \item{sample_qc_flag}{Indicates whether there is a qc error for the sample's metadata (not quality of the actual sample)}
 \item{sample_flag_reason}{Describes why the sample has been flagged. For example, two samples sharing the same sample_location value}
 \item{sample_qc_log}{Log of changes to any sample metadata during qa/qc}

}
}
\keyword{datasets}
