% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_container_inventory}
\alias{sample_container_inventory}
\title{Location information for all sample storage containers}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1361 rows and 15 columns.}
\usage{
sample_container_inventory
}
\description{
A dataset containing location information for the bulk storage of biological samples obtain from fish caught by the Juvenile Salmon Program
}
\keyword{datasets}
