% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sealice_field}
\alias{sealice_field}
\title{Enumeration of sea lice found on fish in the field}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1372 rows and 29 columns.}
\usage{
sealice_field
}
\description{
A dataset containing counts of sea lice (\emph{Caligus clemensi} and \emph{Lepeophtheirus salmonis}) and body abnormalities found on fish collected in the field by the Juvenile Salmon Program
}
\details{
\describe{
 \item{ufn}{The alphanumeric string that uniquely identifies a fish. \strong{Foreign key to}: \code{\link{fish_field_data}}.}
 \item{cal_cope_field}{Field enumeration of \emph{C. clemensi} (Cal.) sea lice, copepodid life stage}
 \item{lep_cope_field}{Field enumeration of \emph{L. salmonis} (Lep.) sea lice, copepodid life stage}
 \item{chal_a_field}{Field enumeration of combined Cal. and Lep. sea lice, chalimus A life stage}
 \item{chal_b_field}{Field enumeration of combined Cal. and Lep. sea lice, chalimus B life stage}
 \item{cal_mot_field}{Field enumeration of motile Cal. sea lice, combined pre-adult and adult life stages and both sexes (non-gravid females only)}
 \item{cgf_field}{Field enumeration of gravid Cal. sea lice}
 \item{lpam_field}{Field enumeration of Lep. sea lice, pre-adult males}
 \item{lpaf_field}{Field enumeration of Lep. sea lice, pre-adult females}
 \item{lam_field}{Field enumeration of Lep. sea lice, adult males}
 \item{laf_field}{Field enumeration of Lep. sea lice, adult females (non-gravid)}
 \item{lgf_field}{Field enumeration of gravid Lep. female sea lice}
 \item{unid_cope_field}{Field enumeration of unidentifiable Cal. or Lep. sea lice, copepodid life stage}
 \item{unid_chal_field}{Field enumeration of unidentifiable Cal. or Lep. sea lice, chalimus A and B life stages}
 \item{chal_scar}{Number of sea louse chalimus scars on the fish's body}
 \item{mot_scar}{Number of motile sea louse scars on the fish's body}
 \item{pred_scar}{Number of predator scars on the fish's body}
 \item{hemorrhaging}{Number of hemorrhaging areas on the fish's body}
 \item{eroded_gill_plate}{Number of eroded gill plates}
 \item{grazed_gill_plate}{Number of grazed gill plates}
 \item{mate_guarding}{Number of sea lice exhibiting mate guarding activity}
 \item{pinched_belly}{Number of occurences of a pinched belly on the fish}
 \item{comments_fish_health_field}{Description of locations of observed body abnormalities and deformities}
 \item{licing_protocol_field}{The protocol that was followed for identifying sea lice in the field. \strong{Categories:}
  \itemize{
   \item \samp{salmoncoast_allstages} = fine-scale identification of all sea lice and life stages
   \item \samp{salmoncoast_motiles} = identification of only motile stages of sea lice
   }}

}
}
\keyword{datasets}
