% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sealice_lab_finescale}
\alias{sealice_lab_finescale}
\title{Identification of sea lice collected from fish in the lab in 2015 from lice collection protocol *lab_fine_2015*}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1486 rows and 33 columns.}
\usage{
sealice_lab_finescale
}
\description{
A dataset containing detailed sea lice taxonomic identification
\describe{
 \item{ufn}{The alphanumeric string that uniquely identifies a fish. \strong{Foreign key to}: \code{\link{fish_lab_data}}.}
 \item{date_liced}{The date the lice were identified.}
 \item{lab_count_fine_no_id}{The number of lice collected that the dissecting technician recorded when the lice were originally collected}
 \item{lab_count_fine_total}{The total number of lice that the louse identifier counted in the sample vial}
 \item{lep_cop}{\emph{Lepeophtheirus salmonis} copepodite.}
 \item{lep_cunifer_cop}{\emph{Lepeophtheirus cunifer} copepodite.}
 \item{lep_chal_a}{\emph{Lepeophtheirus salmonis} chalimus A.}
 \item{lep_chal_b}{\emph{Lepeophtheirus salmonis} chalimus B.}
 \item{lep_pa_m_1}{\emph{Lepeophtheirus salmonis} pre-adult male stage 1.}
 \item{lep_pa_m_2}{\emph{Lepeophtheirus salmonis} pre-adult male stage 2.}
 \item{lep_pa_f_1}{\emph{Lepeophtheirus salmonis} pre-adult female stage 1.}
 \item{lep_pa_f_2}{\emph{Lepeophtheirus salmonis} pre-adult female stage 2.}
 \item{lep_pa_unid}{\emph{Lepeophtheirus salmonis} pre-adult unidentified stage of sex.}
 \item{lep_a_m}{\emph{Lepeophtheirus salmonis} adult male.}
 \item{lep_a_f}{\emph{Lepeophtheirus salmonis} adult female.}
 \item{lep_grav_f}{\emph{Lepeophtheirus salmonis} gravid female.}
 \item{cal_cop}{\emph{Caligus clemensi} copepodite.}
 \item{cal_chal_a_1}{\emph{Caligus clemensi} chalimus A stage 1.}
 \item{cal_chal_a_2}{\emph{Caligus clemensi} chalimus A stage 2.}
 \item{cal_chal_b_3}{\emph{Caligus clemensi} chalimus B stage 3.}
 \item{cal_chal_b_4_f}{\emph{Caligus clemensi} chalimus B stage 4 female.}
 \item{cal_chal_b_4_m}{\emph{Caligus clemensi} chalimus B stage 4 male.}
 \item{cal_chal_4_unid}{\emph{Caligus clemensi} chalimus B stage 4 unidentified sex.}
 \item{cal_chal_a_unid}{\emph{Caligus clemensi} chalimus A unidentified sex.}
 \item{cal_chal_b_unid}{\emph{Caligus clemensi} chalimus B unidentified sex.}
 \item{cal_pa_m}{\emph{Caligus clemensi} pre-adult male.}
 \item{cal_pa_f}{\emph{Caligus clemensi} pre-adult female.}
 \item{cal_grav_f}{\emph{Caligus clemensi} gravid female.}
 \item{cal_mot_unid}{Motile \emph{Caligus clemensi}.}
 \item{unid_louse}{Unidentified louse}
 \item{sealice_comments}{Comments pertaining to sealice ID}
 }
}
\keyword{datasets}
