% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sealice_lab_motiles}
\alias{sealice_lab_motiles}
\title{Enumeration of sea lice found on fish in the lab}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5277 rows and 13 columns.}
\usage{
sealice_lab_motiles
}
\description{
A dataset containing counts of motile sea lice (\emph{Caligus clemensi} and \emph{Lepeophtheirus salmonis}) found on lab-dissected fish collected by the Juvenile Salmon Program
}
\details{
\describe{
 \item{ufn}{The alphanumeric string that uniquely identifies a fish. \strong{Foreign key to}: \code{\link{fish_lab_data}}.}
 \item{cm_lab}{Lab enumeration of \emph{C. clemensi} (Cal.) and \emph{L. salmonis} (Lep.) sea lice, adult and pre-adult males}
 \item{cpaf_lab}{Lab enumeration of Cal. sea lice, pre-adult females}
 \item{caf_lab}{Lab enumeration of Cal. sea lice, adult females (non-gravid)}
 \item{cgf_lab}{Lab enumeration of gravid Cal. sea lice}
 \item{lgf_lab}{Lab enumeration of gravid Lep. sea lice}
 \item{laf_lab}{Lab enumeration of Lep. sea lice, adult females (non-gravid)}
 \item{lpaf_lab}{Lab enumeration of Lep. sea lice, pre-adult females (both first and second stages)}
 \item{lpam_lab}{Lab enumeration of Lep. sea lice, pre-adult males (both first and second stages)}
 \item{lam_lab}{Lab enumeration of Lep. sea lice, adult males}
 \item{ucal_lab}{Lab enumeration Cal. sea lice, unidentifiable sex/life stage}
 \item{ulep_lab}{Lab enumeration Lep. sea lice, unidentifiable sex/life stage}
 \item{lab_count_no_id}{The number of lice that were picked off a fish and put into a formalin sample. This count includes all life stages, species and sex of sea lice. Counting the number of lice like this occured in 2015.}
}
}
\keyword{datasets}
