% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seine_data}
\alias{seine_data}
\title{Time, location, and catch effort of sets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 594 rows and 24 columns.}
\usage{
seine_data
}
\description{
A dataset containing variables collected from Juvenile Salmon Program seines
}
\details{
\describe{
 \item{seine_id}{A unique ID assigned to each seine conducted within a survey. \strong{Primary key to}: \code{\link{fish_field_data}}, \code{\link{bycatch_mort}}.}
 \item{survey_id}{A unique alphanumeric code assigned to every survey. A survey ID may be associated with multiple seines. \strong{Foreign key} to \code{\link{survey_data}}.}
 \item{set_number}{The order in which the seine(s) was/were deployed in a survey}
 \item{set_type}{Denotes whether the set was conducted upon an observed school of fish or in a random area where surface activity wasn't targetted. \samp{NA} means we conducted a visual survey for surface activity but did not set upon any fish.}
 \item{time_searching}{Length of time visually searching for fish before setting the net (only recorded in 2015/2016). \strong{Units:} minutes}
 \item{set_time}{Bunt time, i.e., when the trailing sternward end of the seine is tied onto the boat and a school of fish has been concentrated within the seine. \strong{Format:} hh:mm:ss}
 \item{lat}{Location of fish capture. \strong{Format:} decimal degrees}
 \item{long}{Location of fish capture. \strong{Format:} decimal degrees}
 \item{so_taken}{Number of sockeye salmon retained in a seine.}
 \item{so_total}{Total number of sockeye caught in a seine.}
 \item{pi_taken}{Number of pink salmon retained in a seine.}
 \item{pi_total}{Total number of pink salmon caught in a seine. \samp{NA} = individuals of this species were not enumerated}
 \item{cu_taken}{Number of chum salmon retained in a seine.}
 \item{cu_total}{Total number of chum salmon caught in a seine. \samp{NA} = individuals of this species were not enumerated}
 \item{co_taken}{Number of coho salmon retained in a seine.}
 \item{co_total}{Total number of coho salmon caught in a seine. \samp{NA} = individuals of this species were not enumerated}
 \item{he_taken}{Number of herring retained in a seine.}
 \item{he_total}{Total number of herring caught in a seine. \samp{NA} = individuals of this species were not enumerated}
 \item{ck_taken}{Number of chinook salmon  retained in a seine.}
 \item{ck_total}{Total number of chinook salmon  caught in a seine. \samp{NA} = individuals of this species were not enumerated}
 \item{preservation_status}{\samp{0} = one or more samples from this seine were improperly preserved in the dewar (dewar likely thawed, warming up the samples and degrading their quality). \samp{1} = all samples were properly preserved.}
 \item{fish_retained}{\samp{yes} = at least 1 fish had been retained from the seine. \samp{no} = no fish retained (seine either failed to capture fish, or no fish were taken from a seine)}
 \item{collection_protocol}{The decision-making guidelines that dictate how fish were sampled in the field. \strong{Categories:}
  \itemize{
   \item \samp{SEMSP} = standard Hakai field protocol.
   \samp{DFO} = collection aboard the Nordic Queen purse seiner.
   \samp{SJUBC} = Sam James UBC for diel feeding study fish collection.
   \samp{GM} = Gideon Mordecai for fish that were targeted for collection because they appeared sick.
   }
   }
 \item{seine_comments}{Comments related to the seine}

}
}
\keyword{datasets}
