% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_activity}
\alias{site_activity}
\title{Fish surface activty observed during survey transects}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 699 rows and 5 columns.}
\usage{
site_activity
}
\description{
Before seining, a visual transect 1 nm in length was conducted to quantify the abundance of fish surface activity
}
\details{
\describe{
 \item{survey_id}{A unique alphanumeric code assigned to every survey. \strong{Foreign key} to \code{\link{survey_data}}.}
 \item{school_number}{A sequential number assigned to a discrete group of surface activity that we considered a school of fish}
 \item{school_sliders}{The number of individual surface events that "slid" out the water in this school. \strong{Categories:} \samp{0, 1, 10, 100, 1000}}
 \item{school_poppers}{The number of individual surface events that "popped" out the water in this school. \strong{Categories:} \samp{0, 1, 10, 100, 1000}}
 \item{school_dimpling}{The number of individual surface events that looked like simpling in this school. \strong{Categories:} \samp{0, 1, 10, 100, 1000}}

}
}
\keyword{datasets}
