% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_id}
\alias{stock_id}
\title{Probable genetic stocks associated a fish}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2496 rows and 21 columns.}
\usage{
stock_id
}
\description{
A dataset containing the results of genetic stock ID analyses performed on DNA samples collected from fish caught by the Juvenile Salmon Program.
}
\details{
\describe{
 \item{ufn}{The alphanumeric string that uniquely identifies a fish}
 \item{semsp_id}{A unique ID for every fish retained that was a concatenation of \code{survey_date}, \code{site_id}, \code{species}, and order the fish was pulled out of the net. In 2017 this was method was abandonded, and instead we assigned seine IDs and UFNs to uniquely identify seines and fish}
 \item{sample_id}{The alphanumeric string that uniquely identifies a sample. This enables results to be linked back to the sample and fish from which it was collected. \strong{Format}: \samp{'S'} followed by a sequence of letter and numbers.}
 \item{stock_1}{The genetic stock assigned with the highest probability}
 \item{region_1}{The region associated with Stock 1}
 \item{prob_1}{The probablilty the fish originates from Stock 1}
 \item{stock_2}{The genetic stock assigned with the second highest probability}
 \item{region_2}{The region associated with stock 2}
 \item{prob_2}{The probablilty the fish originates from stock 2}
 \item{stock_3}{The genetic stock assigned with the third highest probability}
 \item{region_3}{The region associated with stock 3}
 \item{prob_3}{The probablilty the fish originates from stock 3}
 \item{stock_4}{The genetic stock assigned with the fourth highest probability}
 \item{region_4}{The region associated with stock 4}
 \item{prob_4}{The probablilty the fish originates from stock 4}
 \item{stock_5}{The genetic stock assigned with the fifth highest probability}
 \item{region_5}{The region associated with stock 5}
 \item{prob_5}{The probablilty the fish originates from stock 5}

}
}
\keyword{datasets}
