% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stomach_metadata}
\alias{stomach_metadata}
\title{Storage information for stomach samples}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4513 rows and 12 columns.}
\usage{
stomach_metadata
}
\description{
Metadata associated with stomach samples obtained from fish caught by the Juvenile Salmon Program, for use in gut content analysis.
}
\details{
\describe{
 \item{ufn}{The alphanumeric string that uniquely identifies a fish. \strong{Foreign key} to \code{\link{fish_lab_data}}. \strong{Format:} \samp{'U'} followed by a sequence of numbers.}
 \item{sample_type}{The type of sample being collected.}
 \item{sample_id}{The alphanumeric string that uniquely identifies a sample. This enables results to be linked back to the sample and fish from which it was collected. \strong{Format}: \samp{'S'} followed by a sequence of letter and numbers.}
 \item{container_id}{The string that uniquely identifies the tray in which sets of sample vials are stored.}
 \item{container_cell}{Vials within a stomach tray are not tracked, so values are \samp{NA}}
 \item{sample_location}{A concatenation of \code{container_id} and \code{container_cell}. Functionally, it is the same as container_id due to cell locations not being tracked for stomach samples.}
 \item{comments_sample}{Comments regarding notable characteristics/properties of the particular sample}
 \item{sample_status}{Results status of the sample}
 \item{sample_qc_flag}{Indicates whether there is a qc error for the sample's metadata (not quality of the actual sample)}
 \item{sample_flag_reason}{Describes why the sample has been flagged.}
 \item{sample_qc_log}{Log of changes to any sample metadata during qa/qc}

}
}
\keyword{datasets}
