% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{Site, date, and other metadata associated with a survey}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 542 rows and 22 columns.}
\usage{
survey_data
}
\description{
A dataset containing variables collected from Juvenile Salmon Program site surveys
}
\details{
\describe{
  \item{survey_id}{A unique alphanumeric code assigned to every survey. \strong{Primary key to}: \code{\link{seine_data}}, \code{\link{site_activity}}, \code{\link{ysi}}, \code{ctd_casts}, and \code{\link{zoop_tows}}.}
  \item{survey_date}{Date upon which a survey was conducted. \strong{Format}: \samp{yyyy-mm-dd}}
  \item{site_id}{A unique identifier given for the geographic location of a survey}
  \item{region}{General geographic zone of a site. \samp{DI} = Discovery Islands, \samp{JS} = Johnstone Strait}
  \item{zone}{Categorization of the generalized migration route taken by outmigrating juvenile salmon (\samp{West}, \samp{Central}, or \samp{East} for Discovery Islands; \samp{N} or \samp{S} for Johnstone Strait).}
  \item{site_name}{Chart name for the location of a survey site}
  \item{crew}{Full names of crew members}
  \item{precip}{Subjective categorical representation of the amount of precipitation. \strong{Categories}: \samp{none, light, moderate, heavy}}
  \item{cloud_cover}{Visual approximation of the \emph{percent} of cloud cover overhead. \strong{Categories}: \samp{0, 25, 50, 75, 100}}
  \item{sea_state}{\strong{Categories} describing surface conditions:
   \itemize{
    \item \samp{0} = Sea-surface smooth and mirror like.
    \item \samp{1} = Scaly ripples, no foam crests.
    \item \samp{2} = Small wavelets, crests glassy, no breaking.
    \item \samp{3} = Large wavelets, crests begin to break, scattered whitecaps.
    \item \samp{4} = small waves 1–4 ft. becoming longer, numerous whiteaps.
    \item \samp{5} = Moderate waves 4–8 ft. taking longer form, many whitecaps, some spray.
    }
    }
  \item{wind_speed}{Estimated wind speed range observed on site. \strong{Units}: knots. \strong{Categories}: \samp{0, 0.1-5, 5-10, 10-15, 15-20, 20-30.}}
  \item{wind_direction}{Approximation of which way the wind is coming from. \strong{Categories}: \samp{N, S, E, W, NE, NW, SE, SW.}}
  \item{tide_state}{Tidal state at the start of the survey. \strong{Categories}: \samp{high slack, low slack, flood, ebb tide.}}
  \item{survey_time_start}{24-hour time of commencing visual survey for site activity. \strong{Format}: \samp{hh:mm:ss}}
  \item{survey_time_end}{24-hour time of end of visual survey for site activity. \strong{Format}: \samp{hh:mm:ss}}
  \item{ctd_cast_id}{An automatically-generated concatenation of \code{site_id}, \code{survey_date}, and drop number. \strong{Primary key} to join \code{\link{survey_data}} to CTD data.}
  \item{zoop_sample_id}{The unique code that identifies a zooplankton sample that is stored in a 250 mL PET jar. \strong{Primary key} to join \code{\link{survey_data}} to \code{\link{zoop_tows}}.}
  \item{secchi}{Secchi depth reading. \strong{Units}: metres (rounded to nearest 0.1 m)}
  \item{net_sets}{The number of times the seine was deployed in a survey}
  \item{set_type}{How (or for whom) the survey was conducted. \samp{Standard} surveys can be used for the purpose of CPUE calculations; other categories (\samp{non-standard} or \samp{DFO} should be excluded.}
  \item{survey_comments}{Field notes specific to survey conditions}
  \item{sampling_week}{A calculated value. Julian day (j) + 4 days for historical categorization purposes, integer divide by 7. \eqn{j + 4 / 7}}

}
}
\keyword{datasets}
