% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ysi}
\alias{ysi}
\title{Temperature and salinity data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 847 rows and 6 columns.}
\usage{
ysi
}
\description{
A dataset containing oceanographic variables collected from Juvenile Salmon Program site surveys with a YSI Pro30 multiparameter meter
}
\details{
\describe{
 \item{survey_id}{A unique alphanumeric code assigned to every survey}
 \item{line_out_depth}{The depth at which the measurement was taken (in metres)}
 \item{temperature}{Temperature in degrees celcius}
 \item{salinity}{Salinity in parts per thousand}
 \item{ysi_gather_lat}{The geopgraphical latitude of measurement gathered by the electronic tablet}
 \item{ysi_gather_long}{The geopgraphical longitude of measurement gathered by the electronic tablet}
}
}
\keyword{datasets}
