% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoop_tows}
\alias{zoop_tows}
\title{Zooplankton tow metadata}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 372 rows and 15 columns.}
\usage{
zoop_tows
}
\description{
A dataset containing variables collected from horizontal zooplankton tows during Juvenile Salmon Program site surveys
}
\details{
\describe{
 \item{survey_id}{A unique alphanumeric code assigned to every survey}
 \item{sample_id}{A unique sample ID assigned and labelled onto every 250 mL PET jar that resulted from the zoop tow}
 \item{zoop_tow_gather_lat}{The geopgraphical latitude of measurement gathered by the electronic tablet}
 \item{zoop_tow_gather_long}{The geopgraphical longitude of measurement gathered by the electronic tablet}
 \item{tow_start_time}{The 24-hour time the zooplankton tow began}
 \item{tow_end_time}{The 24-hour time the zooplankton tow ended}
 \item{tow_duration}{The duration of the zooplakton tow, in minutes}
 \item{flow_meter}{The serial number of the flowmeter}
 \item{flowmeter_in}{The revolution count of the flowmeter before begining the tow}
 \item{flowmeter_out}{The revolution count of the flowmeter after finishing the tow}
 \item{revolutions}{The number of revolutions of the flowmeter that occured during the tow. The difference between \code{\link{flowmeter_in}} and \code{\link{flowmeter_out}}}
 \item{volume}{The volume of water sampled taking in to account on the flowmeter count, the calibration coefficient of the flowmeter, and the size of the net. }
 \item{modelled_volume}{A modelled estimate of the volume sampled. Early season in 2015 the flowmeter was not working, so volume was modeled based on tow duration and secchi depth. The R script of this model is available upon request}
 \item{quality_level}{Raw infers there is no way to check the quality of the data and no guarantee of accuracy is possible, use at your own risk. Technician infers that the data has passed an initial QA/QC level}
 \item{quality_log}{Comments regarding the Quality level}

}
}
\keyword{datasets}
