% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dna_metadata}
\alias{dna_metadata}
\title{Example Title take dna_metadata for example}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2580 rows and 12 columns.}
\usage{
dna_metadata
}
\description{
This is an example description
}
\details{
\describe{
  \item{variable_name}{variable description}
}
}
\keyword{datasets}
