% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_lab_data}
\alias{fish_lab_data}
\title{Biological measurements and metadata associated with fish dissected in the lab}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5259 rows and 17 columns.}
\usage{
fish_lab_data
}
\description{
A dataset containing variables collected from laboratory dissections of fish caught in the field by the Juvenile Salmon Program. A subset of \code{\link{fish_field_data}}
}
\keyword{datasets}
