% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_cumul_abund.R
\name{log_cumul_abund}
\alias{log_cumul_abund}
\title{Predict the cumulative proportion of catch for the purposes of estimating annual migration timing using a logistic growth equation}
\usage{
log_cumul_abund(percent, date)
}
\arguments{
\item{percent}{The cumuative proportion of catch between 0 and 100 percent}

\item{date}{The julian date of capture}
}
\value{
A dataframe with a date column and predicted cumulative proportion of catch
}
\description{
Predict the cumulative proportion of catch for the purposes of estimating annual migration timing using a logistic growth equation
}
