% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_cond.R
\name{mutate_cond}
\alias{mutate_cond}
\title{A condition mutate function}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to mutate one
}
\examples{
mutate_cond()
}
\keyword{conditional}
\keyword{mutate}
