% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_cond.R
\name{mutate_cond}
\alias{mutate_cond}
\title{A conditional mutate function that allows you to mutate a variable based on another}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data, }{variable to test for logical equality, variable you want to mutate}
}
\description{
A conditional mutate function that allows you to mutate a variable based on another
}
\examples{
df <- as.data.frame(x = c(1,2,3), y = c(a,b,c))
mutate_cond(df, x == 1, y == "c" )
}
\keyword{conditional}
\keyword{mutate}
