% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_container_inventory}
\alias{sample_container_inventory}
\title{Location information for all sample storage containers}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1361 rows and 15 columns.}
\usage{
sample_container_inventory
}
\description{
A dataset containing location information for the bulk storage of biological samples obtain from fish caught by the Juvenile Salmon Program
}
\details{
\describe{
 \item{container_id}{The string that uniquely identifies the box, book, tray, etc. in which a sample is stored. \strong{Format:} \samp{B} followed by the two (or three)-character sample type code, and then a number to uniquely identify the box within its sample type (e.g. \samp{"BFA12"}).
  \itemize{
   \item In some instances, a species code may precede the number (where boxes were only unique on a species-level, e.g. \samp{"BFASO1"}).
   \item The only exception is for fin clips, which are stored on Whatman sheets. For samples from 2015-2017, Whatman sheets had the \code{container_id} format \samp{'DW'} or \samp{'JW'} followed by a number, where \samp{'D'} identifies samples collected in the Discovery Islands, and \samp{'J'} denotes Johnstone Strait.
   }
   }
 \item{container_type}{The type of samples being stored in the container}
 \item{storage_medium}{The medium in which a tissue sample is being preserved. \strong{Categories}: \samp{frozen -80, frozen -20, formalin, dry, whatman sheet, 95 percent non-denatured ethanol}}
 \item{current_location}{Physical location or building in which the sample container exists, e.g. \samp{"Quadra Prep Lab"} or \samp{"UBC"}}
 \item{storage_unit}{Identity of the freezer or dry shelving unit in which a sample container is being stored}
 \item{storage_tier_1}{The first organizational tier of storage for the sample container, e.g. samp{"Shelf 1"} or \samp{"Group Box 1"}}
 \item{storage_tier_2}{The second organizational tier of storage for the sample container, within the first organizational tier, e.g. \samp{"Tray 1"}}
 \item{shipping_container}{Number of a box containing a group of sample containers to be shipped from Hakai}
 \item{date_shipped}{Date when a shipment of samples was sent out from Hakai. \strong{Format:} \samp{yyyy-mm-dd}}
 \item{shipped_to}{Destination of the shipping container}
 \item{date_received}{Date when a shipment of samples was received. \strong{Format:} \samp{yyyy-mm-dd}}
 \item{expecting_return}{Whether or not it is expected that shipped samples are to be returned to Hakai}
 \item{date_returned}{Date when processed samples have been returned to Hakai for storage. \strong{Format:} \samp{yyyy-mm-dd}}

}
}
\keyword{datasets}
