% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_colour_hakai}
\alias{scale_colour_hakai}
\title{Hakai scale colour for ggplot2}
\usage{
scale_colour_hakai(theme = "hakai", hakai_key = list(hakai =
  c("#aa1f25", "#78B7C5", "#346257", "#EBCC2A", "#a58457", "#94911c",
  "#6699c5", "#c7e64b")))
}
\description{
hakai scale color
}
\examples{
scale_colour_hakai(theme="hakai")
}
\author{
Brett Johnson
}
