% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_fill_hakai}
\alias{scale_fill_hakai}
\title{Hakai scale fill for ggplot2}
\usage{
scale_fill_hakai(theme = "hakai", hakai_key = list(hakai = c("#aa1f25",
  "#78B7C5", "#346257", "#EBCC2A", "#a58457", "#94911c", "#6699c5",
  "#c7e64b")))
}
\description{
hakai scale fill
}
\examples{
scale_fill_hakai(theme="hakai")
}
\author{
Brett Johnson
}
