% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seine_data}
\alias{seine_data}
\title{Time, location, and catch effort of sets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 594 rows and 24 columns.}
\usage{
seine_data
}
\description{
A dataset containing variables collected from Juvenile Salmon Program seines
}
\details{
\describe{
 \item{seine_id}{Unique ID assigned to each seine conducted}
 \item{survey_id}{A unique survey ID assigned to every survey. A survey ID may be associated with multiple seines.}
 \item{set_number}{The order in which the seine was deployed in a survey. "NA" if no sets were performed}
 \item{set_type}{whether the set was conducted upon an observed school of fish or in a random area where surface activity wasn't targetted. "NA" means we conducted a visual survey for surface activity but did not detect any fish.}
 \item{time_searching}{Only recorded in 2015/2016: length of time visually searching for fish before setting the net.}
 \item{set_time}{Bunt time, i.e., when the trailing sternward end of the seine is tied onto the boat and a school of fish has been concentrated within the seine}
 \item{lat}{Location of fish capture}
 \item{long}{Location of fish capture}
 \item{so_taken}{Number of sockeye salmon retained in a seine.}
 \item{so_total}{Total number of sockeye caught in a seine.}
 \item{pi_taken}{Number of pink salmon retained in a seine.}
 \item{pi_total}{Total number of pink salmon caught in a seine. "NA" = individuals of this species were not enumerated}
 \item{cu_taken}{Number of chum salmon retained in a seine.}
 \item{cu_total}{Total number of chum salmon caught in a seine. "NA" = individuals of this species were not enumerated}
 \item{co_taken}{Number of coho salmon retained in a seine.}
 \item{co_total}{Total number of coho salmon caught in a seine. "NA" = individuals of this species were not enumerated}
 \item{he_taken}{Number of herring retained in a seine.}
 \item{he_total}{Total number of herring caught in a seine. "NA" = individuals of this species were not enumerated}
 \item{ck_taken}{Number of chinook salmon  retained in a seine.}
 \item{ck_total}{Total number of chinook salmon  caught in a seine. "NA" = individuals of this species were not enumerated}
 \item{preservation_status}{"0" = one or more samples from this seine were improperly preserved in the dewar (dewar likely thawed, warming up the samples and degrading their quality). "1" = all samples were properly preserved.}
 \item{fish_retained}{"Yes" = at least 1 fish had been retained from the seine. "No" = no fish retained (seine either failed to capture fish, or no fish were taken from a seine)}
 \item{collection_protocol}{"SEMSP" refers to standard Hakai field protocol. "DFO" refers to collection aboard the Nordic Queen purse seiner. "SJUBC" refers to Sam James UBC for diel feeding study fish collection. "GM" refers to Gideon Mordecai for fish that were targeted for collection because they appeared sick.}
 \item{seine_comments}{Comments related to the seine}

}
}
\keyword{datasets}
